# Optimize groupPodsByServices and hash functions