When a NetworkPolicy rule had ports specified and had a large AddressGroup which has tens of thousands of Pods, it took quite a little CPU time to execute groupPodsByServices in antrea-agent. And if the rule applied to Pods across all Nodes, it would lead to CPU usage burst of the cluster, affecting the healthy of the whole cluster.

This patch optimizes several key functions in the following way:
1. For groupPodsByServices, if the rule has only numbered port, it returns the result with a fast path without iterating all Pods.
2. For groupPodsByServices, reuse the slice to avoid memory reallocations in the loop.
3. For hashServices and hashGroupMemberPod, construct the string with strings.Builder instead of using the spew library.

The improvement according to benchmark:
Before:
|  Test |  Time  | Memory  |  Allocs | 
|---|---|---|---|
| BenchmarkHashGroupMemberPod-4 |                     10907 ns/op |             920 B/op |        47 allocs/op |
| BenchmarkInsert-4 |                                 11006 ns/op    |         920 B/op    |    47 allocs/op |
|BenchmarkHashServices-4               |            13636 ns/op       |     1008 B/op    |      56 allocs/op  |
|BenchmarkGroupPodsByServicesWithNamedPort-4      |   969610858 ns/op    |      81247860 B/op    |  3652470 allocs/op  |
|BenchmarkGroupPodsByServicesWithoutNamedPort-4    |  952734680 ns/op   |       78859480 B/op     | 3552528 allocs/op  |

After:
|  Test |  Time  | Memory  |  Allocs | 
|---|---|---|---|
 |BenchmarkHashGroupMemberPod-4        |              111 ns/op         |       8 B/op        |   1 allocs/op |
 |BenchmarkInsert-4         |                         184 ns/op          |     8 B/op   |        1 allocs/op |
 |BenchmarkHashServices-4             |               472 ns/op       |        56 B/op       |    6 allocs/op |
 |BenchmarkGroupPodsByServicesWithNamedPort-4   |     72067961 ns/op     |     8844037 B/op   |   301967 allocs/op |
 |BenchmarkGroupPodsByServicesWithoutNamedPort-4   |  686 ns/op      |         704 B/op    |      8 allocs/op |

Fixes #1041